更新

+++ GCXDevelopDemo/GXDevelop/NSString+GXDevelop.h
@@ -43,12 +43,12 @@
 /**
  * 添加文字间距
  */
-- (NSMutableAttributedString *)gxAttributeStringWithFont:(UIFont *)font color:(UIColor *)fontColor spacing:(long)spacing;
+- (NSMutableAttributedString *)gxAttributeStringWithFont:(UIFont *)font color:(UIColor *)fontColor spacing:(CGFloat)spacing;
 /**
  * 添加文字间距 行间距
  */
-- (NSMutableAttributedString *)gxAttributeStringWithFont:(UIFont *)font  color:(UIColor *)fontColor spacing:(long)spacing lineSpacing:(CGFloat)linespacing alignment:(NSTextAlignment)alignment;
-- (NSMutableAttributedString *)gxAttributeStringWithFont:(UIFont *)font  color:(UIColor *)fontColor spacing:(long)spacing alignment:(NSTextAlignment)alignment;
+- (NSMutableAttributedString *)gxAttributeStringWithFont:(UIFont *)font  color:(UIColor *)fontColor spacing:(CGFloat)spacing lineSpacing:(CGFloat)linespacing alignment:(NSTextAlignment)alignment;
+- (NSMutableAttributedString *)gxAttributeStringWithFont:(UIFont *)font  color:(UIColor *)fontColor spacing:(CGFloat)spacing alignment:(NSTextAlignment)alignment;
 
 /**
  *将16进制转成 color
